refactor(shell): organize source layout
