+++ packages/shell/src/command/parse.ts
