+++ packages/shell/src/runtime/index.ts
@@ -1,0 +1,2 @@
+export { createShellRuntime } from "./session.ts"
+export type * from "./types.ts"
