--- packages/shell/src/workspace.ts
@@ -1,14 +1,0 @@
-import { cleanWorkspaceMutationPath, cleanWorkspaceShellPath, runWorkspaceInspectionCommand } from "./workspace-shell.ts"
-import { createReadonlyWorkspaceFs, createWritableWorkspaceFs, workspaceMountPoint } from "./workspace-fs.ts"
-
-export type * from "./workspace-types.ts"
-export type { WorkspaceShellFileSystem } from "./workspace-fs.ts"
-
-export {
-  createReadonlyWorkspaceFs,
-  createWritableWorkspaceFs,
-  cleanWorkspaceMutationPath,
-  cleanWorkspaceShellPath,
-  runWorkspaceInspectionCommand,
-  workspaceMountPoint,
-}
