+++ packages/shell/src/workspace/index.ts
@@ -1,0 +1,15 @@
+import { createReadonlyWorkspaceFs, createWritableWorkspaceFs, workspaceMountPoint } from "./filesystem.ts"
+import { runWorkspaceInspectionCommand } from "./inspection.ts"
+import { cleanWorkspaceMutationPath, cleanWorkspaceShellPath } from "./path.ts"
+
+export type * from "./types.ts"
+export type { WorkspaceShellFileSystem } from "./filesystem.ts"
+
+export {
+  createReadonlyWorkspaceFs,
+  createWritableWorkspaceFs,
+  cleanWorkspaceMutationPath,
+  cleanWorkspaceShellPath,
+  runWorkspaceInspectionCommand,
+  workspaceMountPoint,
+}
