+++ packages/shell/src/workspace/types.ts
