+++ packages/shell/test/runtime.test.ts
@@ -10,7 +10,7 @@ import {
   createWritableWorkspaceFs,
   runWorkspaceInspectionCommand,
   workspaceMountPoint,
-} from "../src/workspace.ts"
+} from "../src/workspace/index.ts"
 import { createCloudflareShellProvider } from "../src/providers/cloudflare.ts"
 
 import type {
@@ -27,7 +27,7 @@ import type {
   ShellSearchQuery,
   ShellStat,
   WritableShellWorkspace,
-} from "../src/workspace.ts"
+} from "../src/workspace/index.ts"
 
 // @ts-expect-error workspace contracts belong to @vitehub/shell/workspace.
 import type { ReadonlyShellWorkspace as RootReadonlyShellWorkspace } from "../src/index.ts"
@@ -155,6 +155,24 @@ function createReadonlyRuntime(workspace: ReadonlyShellWorkspace) {
 }
 
 describe("@vitehub/shell just-bash runtime", () => {
+  it("exposes stable public package subpaths", async () => {
+    await expect(import("@vitehub/shell")).resolves.toMatchObject({
+      analyzeShellCommand: expect.any(Function),
+      createShellRuntime: expect.any(Function),
+    })
+    await expect(import("@vitehub/shell/workspace")).resolves.toMatchObject({
+      cleanWorkspaceShellPath: expect.any(Function),
+      createReadonlyWorkspaceFs: expect.any(Function),
+      runWorkspaceInspectionCommand: expect.any(Function),
+    })
+    await expect(import("@vitehub/shell/providers/just-bash")).resolves.toMatchObject({
+      createJustBashProvider: expect.any(Function),
+    })
+    await expect(import("@vitehub/shell/providers/cloudflare")).resolves.toMatchObject({
+      createCloudflareShellProvider: expect.any(Function),
+    })
+  })
+
   it("creates stateful shell sessions with boundary metadata and one-shot exec sugar", async () => {
     const workspace = new MemoryWorkspace({
       "README.md": "# Docs\n",
